Macro tera::try_get_value

macro_rules! try_get_value {
    ($filter_name:expr, $var_name:expr, $ty:ty, $val:expr) => { ... };
}

Helper macro to get real values out of Value while retaining proper errors in filters

Takes 4 args:

the filter name,

the variable name: use "value" if you are using it on the variable the filter is ran on

the expected type

the actual variable

let arr = try_get_value!("first", "value", Vec<Value>, value);
let val = try_get_value!("pluralize", "suffix", String, val.clone());