Crate tera

Tera

Tera is a template engine based on Jinja2 and the Django template language.

See the site for features and to get started.

Macros

compile_templates
Compile templates or exits process

try_get_value
Helper macro to get real values out of Value while retaining proper errors in filters

Structs

Context
The struct that holds the context of a template rendering.

Error
The Error type.

Map
Represents a JSON key/value type.

Number
Represents a JSON number, whether integer or floating point.

Tera
The main point of interaction in this library.

Enums

ErrorKind
The kind of an error.

Value
Represents any valid JSON value.

Functions

escape_html
Escape HTML following OWASP

from_value
Interpret a serde_json::Value as an instance of type T.

to_value
Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

Type Definitions

FilterFn
The filter function type definition

GlobalFn
The global function type definition

Result
Convenient wrapper around std::Result.

TesterFn
The tester function type definition