Crate tera

Tera

Tera is a template engine based on Jinja2 and the Django template language.

See the site for features and to get started.

Modules

helpers
Re-export some helper fns useful to write filters/fns/tests

Macros

try_get_value
Helper macro to get real values out of Value while retaining proper errors in filters

Structs

Context
The struct that holds the context of a template rendering.

Error
The Error type

Map
Re-export Value and other useful things from serde so apps/tools can encode data in Tera types

Number
Re-export Value and other useful things from serde so apps/tools can encode data in Tera types

Tera
The main point of interaction in this library.

Enums

ErrorKind
The kind of an error (non-exhaustive)

Value
Re-export Value and other useful things from serde so apps/tools can encode data in Tera types

Traits

Filter
The filter function type definition

Function
The global function type definition

Test
The tester function type definition

Functions

escape_html
Escape HTML following OWASP

from_value
Re-export Value and other useful things from serde so apps/tools can encode data in Tera types

to_value
Re-export Value and other useful things from serde so apps/tools can encode data in Tera types

Type Definitions

Result
Convenient wrapper around std::Result.